pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spnf"
version = "0.1.0"
description = "Spectral Prefiltering of Neural Fields"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Mustafa Berk Yaldiz" }]
dependencies = [
  "torch>=2.0",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "tqdm>=4.65",
  "imageio>=2.31",
  "numpy>=1.24",
  "scipy>=1.10",
]

[tool.setuptools.packages.find]
include = ["spnf"]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
